blog_project/requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
lxml==4.9.2
outcome==1.2.0
packaging==23.1
psycopg2==2.9.6
PyJWT==2.7.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
